ModSets Save/Load
Modsets can now be saved and loaded.
Mods that are not for warhammer 3 are now marked as such (as I suspect that the file that I use to manage the mods is used for all total war games which is quite insane but why else would there be that parameter. And the launcher that owns the file seems to be game independent)
Improved and fixed some other things...